# internal/apidoc/openapi.yaml
openapi: 3.1.0
info:
  title: MongoDB Investigation Engine API
  description: >
    API for discovering MongoDB collection relationships, tracing documents,
    and detecting orphan records.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/connections:
    get:
      summary: List all connections
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: List of connections with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
    post:
      summary: Create a MongoDB connection
      description: >
        Validates the connection string by connecting to MongoDB before saving.
        Returns 400 if the connection cannot be established.
      tags: [connections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionRequest"
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          description: Invalid request or connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/connections/{id}:
    get:
      summary: Get a connection by ID
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/ConnectionID"
      responses:
        "200":
          description: Connection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a connection
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/ConnectionID"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/connections/{id}/databases:
    get:
      summary: List MongoDB databases
      description: >
        Connects to MongoDB and lists available databases.
        Excludes system databases (admin, local, config).
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/ConnectionID"
      responses:
        "200":
          description: List of database names
          content:
            application/json:
              schema:
                type: object
                properties:
                  databases:
                    type: array
                    items:
                      type: string
                    example: [myapp, analytics, staging]
        "400":
          description: Cannot connect to MongoDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Connection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/connections/{id}/select-db:
    post:
      summary: Select a database for a connection
      description: >
        Sets the active database for a connection. Must be called before
        listing collections or running scans.
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/ConnectionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectDBRequest"
      responses:
        "200":
          description: Database selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Connection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/connections/{id}/collections:
    get:
      summary: List collections in selected database
      description: >
        Returns all collections in the selected database.
        A database must be selected first via the select-db endpoint.
      tags: [connections]
      parameters:
        - $ref: "#/components/parameters/ConnectionID"
      responses:
        "200":
          description: List of collection names
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      type: string
                    example: [users, orders, payments, subscriptions]
        "400":
          description: No database selected or cannot connect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Connection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/relationships:
    get:
      summary: List discovered relationships
      tags: [relationships]
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: Filter by status (suggested, approved, rejected)
          schema:
            type: string
            enum: [suggested, approved, rejected]
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: List of relationships with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
  /api/relationships/discover:
    post:
      summary: Run relationship discovery
      description: >
        Takes a scan ID and discovers relationships by matching candidate field values
        against _id fields of other collections. Returns suggested relationships
        with confidence scores >= 50%.
      tags: [relationships]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scan_id]
              properties:
                scan_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Discovery complete, returns suggested relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relationship"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Discovery failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/relationships/{id}:
    get:
      summary: Get a relationship
      tags: [relationships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Relationship details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relationship"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/relationships/{id}/approve:
    post:
      summary: Approve a suggested relationship
      tags: [relationships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Relationship approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relationship"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/relationships/{id}/reject:
    post:
      summary: Reject a suggested relationship
      tags: [relationships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Relationship rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relationship"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/orphans/detect:
    post:
      summary: Detect orphan records
      description: >
        Scans all approved relationships and finds source field values that
        reference non-existent documents in the target collection.
        Clears previous orphan results for the connection before re-detecting.
      tags: [orphans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection_id]
              properties:
                connection_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Orphan detection complete
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Orphan"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Detection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/orphans:
    get:
      summary: List orphan records
      tags: [orphans]
      parameters:
        - name: connection_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: List of orphans with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
        "400":
          description: Missing connection_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    ConnectionID:
      name: id
      in: path
      required: true
      description: Connection ID (UUID)
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0)
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      required: false
      description: Number of items to return (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    PaginatedResponse:
      type: object
      properties:
        data:
          description: The paginated data array
          type: array
          items: {}
        total:
          description: Total number of items
          type: integer
        offset:
          description: Number of items skipped
          type: integer
        limit:
          description: Number of items per page
          type: integer
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: my-mongo
        connection_string:
          type: string
          example: mongodb://localhost:27017
        database:
          type: string
          example: myapp
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateConnectionRequest:
      type: object
      required: [name, connection_string]
      properties:
        name:
          type: string
          example: my-mongo
        connection_string:
          type: string
          example: mongodb://localhost:27017
    SelectDBRequest:
      type: object
      required: [database]
      properties:
        database:
          type: string
          example: myapp
...
    Relationship:
      type: object
      properties:
        id:
          type: string
          format: uuid
        connection_id:
          type: string
          format: uuid
        source_collection:
          type: string
          example: orders
        source_field:
          type: string
          example: userId
        target_collection:
          type: string
          example: users
        target_field:
          type: string
          example: _id
        confidence:
          type: number
          format: double
          example: 0.987
          description: Match confidence (0.0 to 1.0)
        matched_values:
          type: integer
          example: 987
        sampled_values:
          type: integer
          example: 1000
        status:
          type: string
          enum: [suggested, approved, rejected]
          example: suggested
        explanation:
          type: string
          description: Human-readable explanation of why this relationship was detected
          example: "45/50 values matched (90.0%). Field name 'userId' suggests 'users' collection. Type compatible (ObjectId→ObjectId). Target field 'users._id' has 100% uniqueness."
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
...
    Orphan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        relationship_id:
          type: string
          format: uuid
        source_collection:
          type: string
          example: comments
        source_field:
          type: string
          example: movie_id
        missing_value:
          type: string
          example: "573a1390f29313caabcd416f"
        created_at:
          type: string
          format: date-time